Encryption: Align with 1Password's opcode01 format.
The format was surprisingly already quite close, which I take as pure chance.
Agile Bits' opcode01 format had two significant differences:
 1. A 'magic bytes' format header. This is useful to bail out early when you
    realise you're not decrypting what you thought you were.
 2. Incorporating the IV into the encryption source. This is effectively
    another sanity check, but it does add some mild security like an extended
    salt.

On top of this, it turns out 1Password does an HMAC-SHA256 on the resulting
cipher text (and some other stuff). I could initially not see the purpose of
this, but I now believe it is an attempt to protect against so-called Chosen
Cipher text Attacks (CCA). The attack model would be someone who has access to
your encrypted data, modifies it and then tricks you into decrypting this
incorrect data. It is unlikely but possible that they could learn something
from this. I think it is worth also incorporating this.

There are two other significant outstanding improvements suggested by Agile
Bits:
 1. Using AES in CTR mode instead of CBC.
 2. Using scrypt for key derivation instead of PBKDF2.
The reasoning for not implementing both of these is their lack of availability
across platforms. Both of these are easily available on Linux, so I see no
reason not to move to them.