Switch to using a bitmap to track secure allocation.
This change was made for several motivating reasons. The benefits are:
 - No dynamic allocation required during `free` any more;
 - Free of memory that was never allocated can now be detected;
 - Bookkeeping-based fragmentation is no longer an issue.

There are the following downsides to this change:
 - The allocator is now heavily platform dependent;
 - The allocation algorithm is significantly more complicated.

The trade offs here were seen to be worth it. In particular, this change opens
the way to increase security by detecting malicious attempts to corrupt the
secure heap. Previously it was possible for a malicious colocated process to
force secure freeing of insecure memory without the allocator noticing. The
outcome would be insecure memory in the secure heap which would then be
obliviously used for future (now-insecure) allocations. Following this change
it should be impossible to trick the allocator into adding insecure memory to
its heap. However, note it is still possible for a malicious co-hosted party
to either (a) read secure memory or (b) munlock secure memory allowing it to be
paged.