Move HMAC salt generation into set_mac.
This modifies the MAC logic to still generate the HMAC salt on demand, but only
do it when setting the MAC of an entry. The logic here is that checking the MAC
should be a read-only operation. This also reduces the number of random bytes
we need to generate because we stop regenerating the HMAC salt each time.