Fix potential secure heap corruption/disable.
In the code modified, if the first allocation succeeds but the second fails we
can trigger a read (and free) or uninitialised memory. Should be fine with
respect to the secure heap as it will just disable it, unless by some
coincidence the garbage pointer points at allocated memory. Either way, still
represents a bug.