anticipate and check for an entry list length out of range
It is conceivably possible that an attacker could manually craft an entry list
equal to the maximum we can deal with (UINT_MAX), and then ask to add an entry
to this list. I can't imagine json-c can actually cope with a list this long,
but just in case we survive until this point, reject attempts to add an entry to
such a list. Note that it is still possible to change existing entries or delete
entries from such a list to get back to a state where you can again add entries.