apply -fno-builtin-memset in build system rather than source
Some versions of GCC don't support toggling this within source code. I don't
think there's a technical barrier here, but the developers have simply decided
it's not worth the complexity. To achieve the effect we want, we now simply
apply this flag within the build system to the one file we need it for.