gui: pull out GTK/X11 ops into their own file
The idea here is to create an interface that is used by gui.c that is
independent of whatever graphics toolkit we're using. This interface (defined in
gui.h) can then be implemented for various back ends. The included gtk.c
implements this for GTK 2 and GTK 3. It looks like this will work for the BSDs
but we'll most likely have to use osascript for MacOS.