FIXME: add a secure open_memstream replacement in lib
This new addition to the API is intended to allow callers to avoid storing
sensitive data in temporary (insecure) buffers when reading from a file or pipe.
By asking the library to read from a file descriptor on their behalf, the data
can be directly passed to a secure buffer without an intermediate bounce buffer.