cli: restructure main logic for generic parallelism
We don't actually parallelise any extra commands yet, but the main flow now has
the logic to fork into a number of threads and run any particular action in
parallel. Next step is to split commands into the segments that must be run in
serial and those that can be run in parallel.

Related to Github #9 "Parallelise remaining actions"