gui: fix type confusion resulting in invalid pointer
This function was receiving a void* that was actually a char** but mistakenly
then casting it to a char*. This bug stems from commit
5b6181fa528b98fe6744192ec2506e44d2674901 "gui: use options.space and options.key
directly" though interestingly this was not the only bug in that commit. The
commit in question also started referencing options.space and options.key which
were typically NULL in pw-gui usage.

The result of all of this was sometimes a segfault and sometimes a false
negative for entry lookup. The present commit fixes the type confusion bug while
the other issue was coincidentally fixed in
62818b81710a9e52782eba70b23ddac606a81df4 "gui: cleanup state by using globals
instead of heap-allocated structures".