switch to consistently heap-allocating EVP_CIPHER_CTXs
We were stack-allocating these in libpasswand and doing a mixture of stack- and
heap-allocation in the test suite. As of OpenSSL 1.1.0 the definition of this
struct is no longer exposed in headers are you are expected to always
heap-allocate them.

Note that in the test suite we avoid using EVP_CIPHER_CTX_reset() in a few cases
where it looks like this is appropriate because some of the older versions of
OpenSSL we're supporting don't have this function.