cli: prevent overwriting an entry using 'set'
The 'update' command is now the one and only way to change the value of an
existing entry. As a reminder, the original motivation for this work was that it
was too easy to think you were setting up a new entry but accidentally overwrite
an existing entry you'd forgotten about. You now need to be explicit about this.

Github: closes #18 "separate update action"