fix: memory leak in allocator
passwand_secure_free() was meant to unset bits in the bitmaps in freelist
chunks, but a logic error meant it usually had no effect. That is, the backing
memory was cleared, but the allocation itself was leaked. This was discovered
through a combination of Valgrind and some recent source changes. The moral of
the story is that we should probably be running Valgrind regularly.