cli: fix double free when deleting entries
After introducing the free on exit code (commit
4e307e3d441fff9e35a6147876d1c051c7660d89) we were then freeing an entry twice;
the last one that gets shuffled back one position during a delete and thus
occupies both the last and second-to-last slot. We were also leaking memory
related to the deleted entry. To fix both these issues, we replace the entry in
the last slot with the deleted one.