depend on libscrypt-kdf instead of patching and building Scrypt from source
As of Scrypt 1.3.0, there is an option --enable-libscrypt-kdf to build a
standalone library containing just the key generation function. This is the
piece that we were patching and building Scrypt to expose. So we can now simply
depend on the functionality upstream in Scrypt.

This is already included in the upcoming Ubuntu Focal 20.04 [0], so once Travis
makes these images available we should be able to install it via apt in CI
instead of building from source. I know of no plans to package this in Homebrew
or Macports, so we may be stuck building from source on macOS.

  [0]: https://packages.ubuntu.com/focal/libscrypt-kdf-dev