gui: add more helpful guidance for Catalina users who can't send_text()
MacOS Catalina's accessibility settings don't make much sense to me. I currently
have a keyboard shortcut wired to an Automator flow that calls pw-gui. To get
this to correctly send text to a target program I need to add the *target
program* to the accessibility list of programs allowed to control the computer.
This seems illogical as it's pw-gui (or arguably Automator) that's controlling
the computer. Still, it seems there is no arguing with the fruit vendor.