expose the GTK lock in a header
We need to separate output code into that for X11 and that for Wayland (the
latter does not yet exist). Rather than do this within the same file, it is
cleaner to move output into separate files (x11.c, wayland.c) that each
implement the send_text() function and then choose between them in the build
system. These implementations of send_text() will each need to acquire
gtk_lock, hence we need to expose it to both.