fix race condition in concurrent set/get test
This test case started a long running `get` and then attempted a `set`,
expecting the `set` to fail due to database locking. This is inherently racy
because we do not control the process scheduling of the `get` and the `set`.

This change restructures the test to deliberately stall the `get` on the
password prompt while we try to run the `set`. This should ensure that the `get`
is still running (and has the database locked) when the `set` attempts to access
the database.