CLI: bypass password confirmation when terminal password is from a chain
This addresses point 1 from 11a0474cfb9b07cbea67daefccb3cd172fc4e071.

Commands that set an entry (`generate`, `set`, and `update`) ask for
confirmation of the password to ensure the user did not typo it. When using a
database chain, this had the unfortunate UI effect of asking for confirmation
but requiring the _terminal_ password instead of an earlier chained password.
Apart from being confusing, this confirmation is unnecessary. If the terminal
password originates from an entry in a chained database, we can be reasonably
sure it was not typoed.

This change makes these commands skip password confirmation in this scenario.