add test cases for operating on a corrupted database
Note that `delete` and `update` are _not_ included in the tests for commands not
modifying the existing database. If you think of corruption in a database as
something quantifiable, the corruption that exhibits from entries with
mismatched encryption criteria is affected by each command in the following way:

* `change-main`: equal corruption
* `check`: equal corruption
* `delete`: decreased corruption
* `generate`: increased corruption
* `get`: equal corruption
* `list`: equal corruption
* `set`: increased corruption
* `update`: decreased or equal corruption

We had been planning to allow `get` and `list`, but with this lens it is
observable that all commands are safe to allow except `generate` and `set`.

This is convenient as it is difficult to disallow `update` on a corrupted
database without significant performance degradation (it would no longer be
possible to take an early exit when finding the matching entry).

The `set` variant also demonstrates the very problem that recently caused
database corruption: `set` incorrectly _does_ modify the database if existing
entries fail to decrypt.

Github: #51 “corrupted entries prevent database recovery”