cli: fix: notify command finalisers of decryption failures
Previously it was possible for an entry to fail decryption and for the
subcommand of the CLI to be unaware of this. E.g. if entry 0 and 2 decrypted
fine but entry 1 did not, the subcommand would see entries 0 and 2 but simply
never see entry 1. This led to the counterintuitive behaviour where commands
like `generate` and `set` would add ignore entries that failed decryption and
still proceed with adding a new entry to the database. This could corrupt a
database by either (1) adding an entry that collided with an existing entry in
both space and key or, more problematically, (2) adding a new entry with
differing encryption parameters to the previously resident entries. The second
case would lead to a database that was no longer usable, though technically
still recoverable with some programmatic manipulation.

Github: #51 “corrupted entries prevent database recovery”