relax allocator page size requirement
The memory allocator was designed around a hard constraint of 4KB pages. However
this is not strictly necessary. Some contemporary platforms do not have 4KB
pages, most notably macOS on ARM has 16KB pages. By relaxing this constraint, we
can run on these platforms, albeit at the expense of wasting the remaining
memory (12KB on macOS on ARM) in the page.