prefer static libscrypt-kdf on non-macOS
In the Linux and BSD ecosystems, whether libscrypt-kdf is packaged in the
default operating systems repositories varies. So users are frequently building
it from source. The Scrypt build system produces a static and shared version of
libscrypt-kdf. Linking against the static one avoids later complications if e.g.
it is installed in a non-standard location.