gui: drop main password earlier when exiting with failure
This is a hardening improvement. Its effect is to make the main password
resident in memory for a shorter amount of time. In particular it thwarts a
(admittedly rare and convoluted) scenario where an attacker can block execution
in `show_error` (e.g. on macOS by intercepting executions of `osascript`) and
try to catch the main password in Passwand’s address space.