duplicate password internally when making a key
Casting the originating `char *` password to a `uint8_t *` was technically
Undefined Behaviour. This cannot easily be avoided by changing the type of
`m_t.data` because this is eventually passed to `scrypt_kdf` that expects a
`uint8_t *`. The safest solution seems to be making a copy with the altered
type.