tune allocator spinlock atomics
These changes capture that it is safe to hoist operations _above_ an `unlock()`.

It is curious to me that `atomic_flag_test_and_set_explicit` does not expose
separate ordering parameters for the store and the load. I am not sure it would
help us in this scenario, especially as we have what amounts to a full fence
following this, but nevertheless it is surprising.